fix: collaboration playlist in playlist handler

+++ src/api/collaborations/handler.js
@@ -1,8 +1,13 @@
 const autoBind = require('auto-bind');
 
 class CollaborationsHandler {
-  constructor(service, validator) {
-    this._service = service;
+  constructor(
+    playlistsService,
+    collaborationsService,
+    validator,
+  ) {
+    this._service = collaborationsService;
+    this._playlistService = playlistsService;
     this._validator = validator;
 
     autoBind(this);
@@ -12,10 +17,12 @@ class CollaborationsHandler {
     this._validator.validatePostCollaborationPayload(request.payload);
 
     const { playlistId, userId } = request.payload;
+    const { id: ownerId } = request.auth.credentials;
+    await this._playlistService.verifyPlaylistOwner({ playlistId, ownerId });
     const collaborationId = await this._service.addCollaboration({ playlistId, userId });
     const response = h.response({
       status: 'success',
-      message: 'Authentication berhasil ditambahkan',
+      message: 'Collaboration berhasil ditambahkan',
       data: {
         collaborationId,
       },
@@ -28,10 +35,12 @@ class CollaborationsHandler {
     this._validator.validateDeleteCollaborationPayload(request.payload);
 
     const { playlistId, userId } = request.payload;
+    const { id: ownerId } = request.auth.credentials;
+    await this._playlistService.verifyPlaylistOwner({ playlistId, ownerId });
     await this._service.deleteCollaboration({ playlistId, userId });
     return {
       status: 'success',
-      message: 'Refresh token berhasil dihapus',
+      message: 'Collaboration berhasil dihapus',
     };
   }
 }
